requirements.txt
Pillow
Flask
pika
